fix changewallet sheet glitch while opening
